Cleanups for v1.0.0! Simplify code and reorganize project
