+++ TLYShyNavBar/Categories/UIScrollView+Helpers.h
@@ -1,0 +1,16 @@
+//
+//  UIScrollView+Helpers.h
+//  TLYShyNavBarDemo
+//
+//  Created by Mazyad Alabduljaleel on 11/13/15.
+//  Copyright © 2015 Telly, Inc. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+
+@interface UIScrollView (Helpers)
+
+- (void)tly_smartSetInsets:(UIEdgeInsets)contentAndScrollIndicatorInsets;
+
+@end
