+++ TLYShyNavBar/ShyControllers/TLYShyStatusBarController.h
@@ -1,0 +1,17 @@
+//
+//  TLYShyStatusBarController.h
+//  TLYShyNavBarDemo
+//
+//  Created by Mazyad Alabduljaleel on 11/13/15.
+//  Copyright © 2015 Telly, Inc. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import "TLYShyParent.h"
+
+
+@interface TLYShyStatusBarController : NSObject <TLYShyParent>
+
+@property (nonatomic, weak) UIViewController *viewController;
+
+@end
