+++ TLYShyNavBar/TLYShyNavBarFade.h
@@ -1,0 +1,21 @@
+//
+//  TLYShyNavBarFade.h
+//  TLYShyNavBarDemo
+//
+//  Created by Mazyad Alabduljaleel on 11/13/15.
+//  Copyright © 2015 Telly, Inc. All rights reserved.
+//
+
+#ifndef TLYShyNavBarFade_h
+#define TLYShyNavBarFade_h
+
+/** This enum helps control the navigation bar fade behavior.
+ */
+typedef NS_ENUM(NSInteger, TLYShyNavBarFade) {
+    
+    TLYShyNavBarFadeDisabled,
+    TLYShyNavBarFadeSubviews,
+    TLYShyNavBarFadeNavbar,
+};
+
+#endif /* TLYShyNavBarFade_h */
